Complete video section

+++ src/utils/cloudinary.js
@@ -8,21 +8,30 @@ cloudinary.config({
 });
 
 const uploadOnCloudinary = async (localFilePath) => {
-    try{
-        if (!localFilePath) return null
-        // upload the file on cloudinary
-      const response = await cloudinary.uploader.upload(localFilePath, {
-            resource_type: "auto"
-        })
-        // file has been uploaded successfull
-        // console.log("file is uploaded on cloudinary", response.url);
-       
-        fs.unlinkSync(localFilePath)
+    try {
+        if (!localFilePath) return null;
+
+        console.log("Uploading file:", localFilePath);
+
+        const response = await cloudinary.uploader.upload(localFilePath, {
+            resource_type: "auto",
+        });
+
+        console.log("Cloudinary Success:", response);
+
+        fs.unlinkSync(localFilePath);
+
         return response;
 
-    } catch (error){
-         fs.unlinkSync(localFilePath) 
-         return null;
+    } catch (error) {
+
+        console.log("Cloudinary Error:", error);
+
+        if (localFilePath && fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        }
+
+        return null;
     }
 }
